Remove all Emacswiki packages
The Emacswiki is completely unsuitable as a source of automatically
generated and widely distributed packages.  Anyone can edit any
package on the Emacswiki without even having to create an account.

This is a huge security risk for the whole community.  And addressing
that is much more important than to avoid inconveniencing some users
by removing their favorite packages.